+++ src/kerko/__init__.py
@@ -56,6 +56,7 @@ def init_default_config(state):
     state.app.config.setdefault('KERKO_RESULTS_ATTACHMENT_LINKS', True)
     state.app.config.setdefault('KERKO_RESULTS_URL_LINKS', True)
     state.app.config.setdefault('KERKO_FACET_COLLAPSING', False)
+    state.app.config.setdefault('KERKO_FULLTEXT_SEARCH', True)
     state.app.config.setdefault('KERKO_PRINT_ITEM_LINK', False)
     state.app.config.setdefault('KERKO_PRINT_CITATIONS_LINK', False)
     state.app.config.setdefault('KERKO_PRINT_CITATIONS_MAX_COUNT', 0)